Separation of diffList & songList ++ live favourites
